first try getting kmTotal and timeTotal for order when carrier is selected

+++ TMS/Classes/City.cs
@@ -9,12 +9,12 @@ namespace TMS.Classes
     internal class City
     {
         public string city;
-        public double distance;//distance is to east city
+        public int distance;//distance is to east city
         public double time;//time is to east city
         public string westCity;
         public string eastCity;
 
-        public City(string city, double distance, double time, string westCity, string eastCity)
+        public City(string city, int distance, double time, string westCity, string eastCity)
         {
             this.city = city;
             this.distance = distance;
